Abstract sandbox functions to a lib
